Use Rhino's require implementation.
