+++ rhino_modules/jsdoc/src/scanner.js
@@ -6,64 +6,63 @@
 	@license Apache License 2.0 - See file 'LICENSE.md' in this project.
  */
 
-(function() {
-	var common = {
-		mixin: require('common/util').mixin,
-		events: require('common/events')
-	};
+
+var common = {
+	mixin: require('common/util').mixin,
+	events: require('common/events')
+};
+
+var fs = require('fs');
+
+/**
+    @constructor
+    @mixes module:common.events
+ */
+exports.Scanner = function() {
+}
+common.mixin(exports.Scanner.prototype, common.events);
+
+/**
+	Recursively searches the given searchPaths for js files.
+	@param {Array.<string>} searchPaths
+	@param {number} [depth=1]
+	@fires sourceFileFound
+ */
+exports.Scanner.prototype.scan = function(searchPaths, depth, includeMatch, excludeMatch) {
+	var filePaths = [],
+	    that = this;
+
+	searchPaths = searchPaths || [];
+	depth = depth || 1;
+
+	searchPaths.forEach(function($) {
+        if ( fs.stat($).isFile() ) {
+            filePaths.push($);
+        }
+        else {
+		    filePaths = filePaths.concat(fs.ls($, depth));
+		}
+	});
 	
-	var fs = require('fs');
+	filePaths = filePaths.filter(function($) {
+	    if (includeMatch && !includeMatch.test($)) {
+	        return false
+	    }
+	    
+	    if (excludeMatch && excludeMatch.test($)) {
+	        return false
+	    }
+	    
+	    return true;
+	});
 	
-	/**
-	    @constructor
-	    @mixes module:common.events
-	 */
-	exports.Scanner = function() {
-	}
-	common.mixin(exports.Scanner.prototype, common.events);
+	filePaths = filePaths.filter(function($) {
+	    var e = { fileName: $ };
+        that.fire('sourceFileFound', e);
+	    
+	    return !e.defaultPrevented;
+	});
 
-	/**
-		Recursively searches the given searchPaths for js files.
-		@param {Array.<string>} searchPaths
-		@param {number} [depth=1]
-		@fires sourceFileFound
-	 */
-	exports.Scanner.prototype.scan = function(searchPaths, depth, includeMatch, excludeMatch) {
-		var filePaths = [],
-		    that = this;
+	return filePaths;
+}
 
-		searchPaths = searchPaths || [];
-		depth = depth || 1;
-
-		searchPaths.forEach(function($) {
-            if ( fs.stat($).isFile() ) {
-                filePaths.push($);
-            }
-            else {
-			    filePaths = filePaths.concat(fs.ls($, depth));
-			}
-		});
-		
-		filePaths = filePaths.filter(function($) {
-		    if (includeMatch && !includeMatch.test($)) {
-		        return false
-		    }
-		    
-		    if (excludeMatch && excludeMatch.test($)) {
-		        return false
-		    }
-		    
-		    return true;
-		});
-		
-		filePaths = filePaths.filter(function($) {
-		    var e = { fileName: $ };
-            that.fire('sourceFileFound', e);
-		    
-		    return !e.defaultPrevented;
-		});
-
-		return filePaths;
-	}
-	
-})();
