+++ rhino_modules/jsdoc/tag/validator.js
@@ -5,48 +5,47 @@
 	@author Michael Mathews <micmath@gmail.com>
 	@license Apache License 2.0 - See file 'LICENSE.md' in this project.
  */
-(function() {
-	
-	var dictionary = require('jsdoc/tag/dictionary');
-	
-	/**
-	    Validate the given tag.
-	 */
-	exports.validate = function(tag, meta) {
-	    var tagDef = dictionary.lookUp(tag.title);
-	    
-	    if (!tagDef && !env.conf.tags.allowUnknownTags) {
-	        throw new UnknownTagError(tag.title, meta);
-	    }
-	    
-	    if (!tag.text) {
-	        if (tagDef.mustHaveValue) {
-	            throw new TagValueRequiredError(tag.title, meta);
-	        }
-	    }
-	    else {
-	        if (tagDef.mustNotHaveValue) {
-	            throw new TagValueNotPermittedError(tag.title, meta);
-	        }
-	    }
-	}
-	
-	function UnknownTagError(tagName, meta) {
-        this.name = 'UnknownTagError';
-        this.message = 'The @' + tagName + ' tag is not a known tag. File: ' + meta.filename + ', Line: ' + meta.lineno + '\n' + meta.comment;
+
+
+var dictionary = require('jsdoc/tag/dictionary');
+
+/**
+    Validate the given tag.
+ */
+exports.validate = function(tag, meta) {
+    var tagDef = dictionary.lookUp(tag.title);
+    
+    if (!tagDef && !env.conf.tags.allowUnknownTags) {
+        throw new UnknownTagError(tag.title, meta);
     }
-    UnknownTagError.prototype = Error.prototype;
-	
-	function TagValueRequiredError(tagName, meta) {
-        this.name = 'TagValueRequiredError';
-        this.message = 'The @' + tagName + ' tag requires a value. File: ' + meta.filename + ', Line: ' + meta.lineno + '\n' + meta.comment;
+    
+    if (!tag.text) {
+        if (tagDef.mustHaveValue) {
+            throw new TagValueRequiredError(tag.title, meta);
+        }
     }
-    TagValueRequiredError.prototype = Error.prototype;
-	
-	function TagValueNotPermittedError(tagName, message, meta) {
-        this.name = 'TagValueNotPermittedError';
-        this.message = 'The @' + tagName + ' tag does not permit a value: "' + message + '". File: ' + meta.filename + ', Line: ' + meta.lineno + '\n' + meta.comment;
+    else {
+        if (tagDef.mustNotHaveValue) {
+            throw new TagValueNotPermittedError(tag.title, meta);
+        }
     }
-    TagValueNotPermittedError.prototype = Error.prototype;
-    	
-})();
+}
+
+function UnknownTagError(tagName, meta) {
+    this.name = 'UnknownTagError';
+    this.message = 'The @' + tagName + ' tag is not a known tag. File: ' + meta.filename + ', Line: ' + meta.lineno + '\n' + meta.comment;
+}
+UnknownTagError.prototype = Error.prototype;
+
+function TagValueRequiredError(tagName, meta) {
+    this.name = 'TagValueRequiredError';
+    this.message = 'The @' + tagName + ' tag requires a value. File: ' + meta.filename + ', Line: ' + meta.lineno + '\n' + meta.comment;
+}
+TagValueRequiredError.prototype = Error.prototype;
+
+function TagValueNotPermittedError(tagName, message, meta) {
+    this.name = 'TagValueNotPermittedError';
+    this.message = 'The @' + tagName + ' tag does not permit a value: "' + message + '". File: ' + meta.filename + ', Line: ' + meta.lineno + '\n' + meta.comment;
+}
+TagValueNotPermittedError.prototype = Error.prototype;
+
