+++ rhino_modules/sys.js
@@ -1,0 +1,4 @@
+
+exports.puts = function(str) {
+    print(String(str));
+};
