--- rhino_modules/sys/index.js
@@ -1,6 +1,0 @@
-
-module.exports = {
-    'puts' : function(str) {
-        print(String(str));
-    }
-};
